CaptchaNet/requirements.txt
torch==1.9.0
torchvision==0.10.0
pytorch-lightning==1.4.2
Pillow==8.2.0
numpy==1.21.0
